+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `cwd` keyword argument in `shell_run_command` - [#41](https://github.com/PrefectHQ/prefect-shell/pull/41)
 
 ### Changed
+- Have `shell_run_command` default to `shell="powershell" if sys.platform == "win32" else "bash"` - [#47](https://github.com/PrefectHQ/prefect-shell/pull/47)
 
 ### Deprecated
 
@@ -20,7 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
-
 
 ## 0.1.2